Inherit from GraphQLObject in the Interface type
